Move main function to main.py
Use a minimal __main__.py as recommended by python.
